+++ test/Catch2Framework.test.ts
@@ -6,6 +6,7 @@ import { inspect } from 'util';
 import * as sinon from 'sinon';
 import { EOL } from 'os';
 import { example1 } from './example1';
+import * as pathlib from 'path';
 import {
   TestAdapter,
   Imitation,
@@ -208,9 +209,11 @@ describe(path.basename(__filename), function () {
         type: 'suite',
         state: 'completed',
         suite: adapter.suite1,
-        description: './ (0ms)',
+        description: `.${pathlib.sep} (0ms)`,
         tooltip:
-          'Name: execPath1.exe\nDescription: ./\n\nPath: <masked>\nCwd: <masked>\n\nTests: 1\n  - failed: 1\n⏱Duration: 0ms',
+          'Name: execPath1.exe\nDescription: .' +
+          pathlib.sep +
+          '\n\nPath: <masked>\nCwd: <masked>\n\nTests: 1\n  - failed: 1\n⏱Duration: 0ms',
       },
       { type: 'finished' },
       { type: 'started', tests: [s1t1.id] },
@@ -229,9 +232,11 @@ describe(path.basename(__filename), function () {
         type: 'suite',
         state: 'completed',
         suite: adapter.suite1,
-        description: './ (0ms)',
+        description: `.${pathlib.sep} (0ms)`,
         tooltip:
-          'Name: execPath1.exe\nDescription: ./\n\nPath: <masked>\nCwd: <masked>\n\nTests: 2\n  - passed: 1\n⏱Duration: 0ms',
+          'Name: execPath1.exe\nDescription: .' +
+          pathlib.sep +
+          '\n\nPath: <masked>\nCwd: <masked>\n\nTests: 2\n  - passed: 1\n⏱Duration: 0ms',
       },
       { type: 'finished' },
     ]);